+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Derive Clone and Debug for AppDirs and UserDirs
+
 ## [0.1.1] - 2019-06-08
 
 ### Fixed